requirements.txt
websockets>=12.0
httpx>=0.25
python-dotenv>=1.0
urllib3>=2.0
rich>=13.0
